LayoutTests/fast/dom/XMLSerializer-serializeToString-entities-expected.txt

Tests that XMLSerializer.serializeToString() properly encodes entities in attribute values when the Element's associated document is an HTMLDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new XMLSerializer()).serializeToString(node) is "<test xmlns=\"http://www.w3.org/1999/xhtml\" a=\"&lt;potato/&gt;\"></test>"
PASS successfullyParsed is true

TEST COMPLETE
